refactor(BA-6176): split role_permission_preset DTOs into their own module
- Move the child-row types (Node, Entry, FailureInfo, BulkAdd/Remove
  inputs and payloads) from ``v2/role_preset/`` into a new
  ``v2/role_permission_preset/`` module so each table owns its own
  surface.
- Add ``RolePermissionPresetFilter`` (with ``role_preset_id``,
  ``entity_type``, ``operation``, ``created_at`` and AND/OR/NOT) plus
  ``RolePermissionPresetOrder`` / ``RolePermissionPresetOrderField`` for
  the upcoming ``RolePresetNode.permissions`` connection resolver.
- Stop re-exporting unrelated common/rbac types through these modules'
  ``types.py`` files; callers import ``OrderDirection`` and the RBAC
  filter/enum types directly from their owning modules.
