+++ src/ai/backend/common/dto/manager/v2/role_preset/__init__.py
@@ -3,10 +3,8 @@ Role Preset DTO v2 models for Manager API.
 """
 
 from ai.backend.common.dto.manager.v2.role_preset.request import (
-    BulkAddRolePermissionPresetsInput,
     BulkDeleteRolePresetsInput,
     BulkPurgeRolePresetsInput,
-    BulkRemoveRolePermissionPresetsInput,
     BulkRestoreRolePresetsInput,
     CreateRolePresetInput,
     RolePresetFilter,
@@ -15,41 +13,23 @@ from ai.backend.common.dto.manager.v2.role_preset.request import (
     UpdateRolePresetInput,
 )
 from ai.backend.common.dto.manager.v2.role_preset.response import (
-    BulkAddRolePermissionPresetsPayload,
     BulkDeleteRolePresetsPayload,
     BulkPurgeRolePresetsPayload,
-    BulkRemoveRolePermissionPresetsPayload,
     BulkRestoreRolePresetsPayload,
-    BulkRolePermissionPresetFailureInfo,
     BulkRolePresetFailureInfo,
     CreateRolePresetPayload,
-    RolePermissionPresetNode,
     RolePresetNode,
     SearchRolePresetsPayload,
     UpdateRolePresetPayload,
 )
-from ai.backend.common.dto.manager.v2.role_preset.types import (
-    EntityType,
-    OperationTypeDTO,
-    OrderDirection,
-    RBACElementTypeDTO,
-    RolePermissionPresetEntry,
-    RolePresetOrderField,
-)
+from ai.backend.common.dto.manager.v2.role_preset.types import RolePresetOrderField
 
 __all__ = (
     # Types
-    "EntityType",
-    "OperationTypeDTO",
-    "OrderDirection",
-    "RBACElementTypeDTO",
-    "RolePermissionPresetEntry",
     "RolePresetOrderField",
     # Request DTOs
-    "BulkAddRolePermissionPresetsInput",
     "BulkDeleteRolePresetsInput",
     "BulkPurgeRolePresetsInput",
-    "BulkRemoveRolePermissionPresetsInput",
     "BulkRestoreRolePresetsInput",
     "CreateRolePresetInput",
     "RolePresetFilter",
@@ -57,15 +37,11 @@ __all__ = (
     "SearchRolePresetsInput",
     "UpdateRolePresetInput",
     # Response DTOs
-    "BulkAddRolePermissionPresetsPayload",
     "BulkDeleteRolePresetsPayload",
     "BulkPurgeRolePresetsPayload",
-    "BulkRemoveRolePermissionPresetsPayload",
     "BulkRestoreRolePresetsPayload",
-    "BulkRolePermissionPresetFailureInfo",
     "BulkRolePresetFailureInfo",
     "CreateRolePresetPayload",
-    "RolePermissionPresetNode",
     "RolePresetNode",
     "SearchRolePresetsPayload",
     "UpdateRolePresetPayload",
