+++ src/ai/backend/common/dto/manager/v2/role_preset/response.py
@@ -7,38 +7,19 @@ from datetime import datetime
 from pydantic import Field
 
 from ai.backend.common.api_handlers import BaseResponseModel
-from ai.backend.common.dto.manager.v2.role_preset.types import (
-    EntityType,
-    OperationTypeDTO,
-    RBACElementTypeDTO,
-)
-from ai.backend.common.identifier.role_permission_preset import RolePermissionPresetID
+from ai.backend.common.dto.manager.v2.rbac.types import RBACElementTypeDTO
 from ai.backend.common.identifier.role_preset import RolePresetID
 
 __all__ = (
-    "BulkAddRolePermissionPresetsPayload",
     "BulkDeleteRolePresetsPayload",
     "BulkPurgeRolePresetsPayload",
-    "BulkRemoveRolePermissionPresetsPayload",
     "BulkRestoreRolePresetsPayload",
-    "BulkRolePermissionPresetFailureInfo",
     "BulkRolePresetFailureInfo",
     "CreateRolePresetPayload",
-    "RolePermissionPresetNode",
     "RolePresetNode",
     "SearchRolePresetsPayload",
     "UpdateRolePresetPayload",
 )
-
-
-class RolePermissionPresetNode(BaseResponseModel):
-    """Node model for a stored permission entry under a role preset."""
-
-    id: RolePermissionPresetID = Field(description="Permission entry UUID.")
-    role_preset_id: RolePresetID = Field(description="UUID of the parent role preset.")
-    entity_type: EntityType = Field(description="Entity type the permission applies to.")
-    operation: OperationTypeDTO = Field(description="Operation granted by the permission.")
-    created_at: datetime = Field(description="Creation timestamp.")
 
 
 class RolePresetNode(BaseResponseModel):
@@ -88,15 +69,6 @@ class BulkRolePresetFailureInfo(BaseResponseModel):
     message: str = Field(description="Error message describing the failure.")
 
 
-class BulkRolePermissionPresetFailureInfo(BaseResponseModel):
-    """Failure detail for a single role-permission-preset row in a bulk operation."""
-
-    permission_preset_id: RolePermissionPresetID = Field(
-        description="role_permission_presets row ID that the operation failed on.",
-    )
-    message: str = Field(description="Error message describing the failure.")
-
-
 class BulkDeleteRolePresetsPayload(BaseResponseModel):
     """Payload for bulk-soft-deleting role presets."""
 
@@ -133,27 +105,6 @@ class BulkPurgeRolePresetsPayload(BaseResponseModel):
     failed: list[BulkRolePresetFailureInfo] = Field(
         default_factory=list,
         description="Role preset IDs that failed to purge.",
-    )
-
-
-class BulkAddRolePermissionPresetsPayload(BaseResponseModel):
-    """Payload for bulk-adding permission entries to a role preset."""
-
-    permissions: list[RolePermissionPresetNode] = Field(
-        description="Permission entries that were added.",
-    )
-
-
-class BulkRemoveRolePermissionPresetsPayload(BaseResponseModel):
-    """Payload for bulk-removing permission entries from a role preset."""
-
-    items: list[RolePermissionPresetNode] = Field(
-        default_factory=list,
-        description="Permission entries that were removed.",
-    )
-    failed: list[BulkRolePermissionPresetFailureInfo] = Field(
-        default_factory=list,
-        description="Permission entry IDs that failed to delete.",
     )
