+++ src/ai/backend/common/dto/manager/v2/role_preset/types.py
@@ -4,24 +4,7 @@ from __future__ import annotations
 
 from enum import StrEnum
 
-from pydantic import Field
-
-from ai.backend.common.api_handlers import BaseRequestModel
-from ai.backend.common.dto.manager.v2.common import OrderDirection
-from ai.backend.common.dto.manager.v2.rbac.types import (
-    EntityType,
-    OperationTypeDTO,
-    RBACElementTypeDTO,
-)
-
-__all__ = (
-    "EntityType",
-    "OperationTypeDTO",
-    "OrderDirection",
-    "RBACElementTypeDTO",
-    "RolePermissionPresetEntry",
-    "RolePresetOrderField",
-)
+__all__ = ("RolePresetOrderField",)
 
 
 class RolePresetOrderField(StrEnum):
@@ -31,10 +14,3 @@ class RolePresetOrderField(StrEnum):
     SCOPE_TYPE = "scope_type"
     CREATED_AT = "created_at"
     UPDATED_AT = "updated_at"
-
-
-class RolePermissionPresetEntry(BaseRequestModel):
-    """A single (entity_type, operation) pair carried by a role preset."""
-
-    entity_type: EntityType = Field(description="Entity type the permission applies to.")
-    operation: OperationTypeDTO = Field(description="Operation granted by the permission.")
